backend/requirements.txt
fastapi==0.88.0
uvicorn==0.24.0
sqlmodel==0.0.11
python-dotenv==1.0.0
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
pydantic<2.0.0,>=1.9.0
thefuzz==0.20.0
rapidfuzz>=3.0.0
tenacity==8.2.3